bgreydon/bugfix and little reworks
s21_sprintf.c fix output that jump in memory incorrect
in any files little changes

+++ src/s21_string/s21_string.h
@@ -3,7 +3,7 @@
 
 #ifndef S21_SIZE_T
 #define S21_SIZE_T
-typedef unsigned int s21_size_t;
+typedef unsigned long s21_size_t;
 #endif  // S21_SIZE_T
 #ifndef S21_NULL
 #define S21_NULL ((void *)0)
